NXP-28995: extract content utilities in tests
